Centralize engine registration & refactor params
Centralize engine registration, refactor parameter registration, and apply small DSP/type fixes.

- Replace per-engine REGISTER_ENGINE stubs with a note that registration is centralized in XOmnibusProcessor.cpp (removes duplicate registration macros from many engine .cpp stubs).
- Convert parameter factories (Ocelot, Owlfish) to a vector-based registration API: addParameters(vector<unique_ptr<RangedAudioParameter>>&), with a createParameterLayout() wrapper for compatibility. Replace many parameter factory helpers with compact F/I/C lambdas that push parameters into the vector.
- Add PresetManager::addPreset(PresetData) to allow programmatic/test insertion of presets.
- Rename/conflict fixes: DCBlocker -> OracleDCBlocker / OuroborosDCBlocker to avoid name collisions and update usages; rename ModalResonator -> OspreyResonator; change Origami noiseState from float to uint32_t for correct type; include FastMath in Overworld header.
- CMake: add Tests/ExportTests/XPNExportTests.cpp to the test target and link juce_gui_basics for test build.

These changes simplify centralized engine handling, enable a shared APVTS parameter registration approach, and fix a few type/naming issues.

+++ Source/Engines/Origami/OrigamiEngine.h
@@ -231,7 +231,7 @@ struct OrigamiVoice
     // Oscillator state
     float sawPhase = 0.0f;
     float sqrPhase = 0.0f;
-    float noiseState = 12345u;
+    uint32_t noiseState = 12345u;
 
     // Glide
     float currentFreq = 440.0f;
